# Replace custom LU decomposition with faster implementation