## Merge Request - GuideLine Checklist 

**Benchmark: Sparse LU Decomposition (Old vs. New vs. MUMPS)**

This PR aims to improve the setup times of the DirectSolver on the coarsest mesh.
It also improves the setup times of the smoothers across-origin boundary matrix.

| Mesh Size | Old Time (s) | New Time (s) | Speedup (Old→New) | MUMPS Time (s) | New / MUMPS |
|-------------|--------------|--------------|-------------------|----------------|-------------|
| 17 × 16     | 0.0074       | 0.0014       | 5.3×              | 0.0023         | 0.6×       |
| 33 × 32     | 0.072        | 0.010        | 7.2×              | 0.0078         | 1.3×       |
| 65 × 64     | 1.25         | 0.078         | 16.0×             | 0.034          | 2.3×       |
| 129 × 128   | 18.90        | 0.78         | 24.2×             | 0.094          | 8.3×      |


**Guideline** to check code before resolve WIP and approval, respectively.
As many checkboxes as possible should be ticked.

### Checks by code author:
Always to be checked:
* [ ] There is at least one issue associated with the pull request.
* [ ] New code adheres with the [coding guidelines](https://github.com/SciCompMod/GMGPolar/wiki)
* [ ] No large data files have been added to the repository. Maximum size for files should be of the order of KB not MB. In particular avoid adding of pdf, word, or other files that cannot be change-tracked correctly by git.

If functions were changed or functionality was added:
* [ ] Tests for new functionality has been added
* [ ] A local test was succesful

If new functionality was added:
* [ ] There is appropriate **documentation** of your work. (use doxygen style comments)

If new third party software is used:
* [ ] Did you pay attention to its license? Please remember to add it to the wiki after successful merging.

If new mathematical methods or epidemiological terms are used:
* [ ] Are new methods referenced? Did you provide further documentation?

### Checks by code reviewer(s):
* [ ] Is the code clean of development artifacts e.g., unnecessary comments, prints, ...
* [ ] The ticket goals for each associated issue are reached or problems are clearly addressed (i.e., a new issue was introduced).
* [ ] There are appropriate **unit tests** and they pass.
* [ ] The git history is clean and linearized for the merge request. All reviewers should squash commits and write a simple and meaningful commit message.
* [ ] Coverage report for new code is acceptable. 
- [ ] * [ ] No large data files have been added to the repository. Maximum size for files should be of the order of KB not MB. In particular avoid adding of pdf, word, or other files that cannot be change-tracked correctly by git.