@julianlitz 2025-08-28
@EmilyBourne You don't need to review this if you don't want to.
I spend multiple days trying to get a working version and the code is quite complex. Since all tests are passing, I assume that it works correctly.

If you ever find that the setup time of the Smoother or the DirectSolver is too slow, then you can revisit this PR and replace the current unoptimized version with this one.